fix: manually running custom homepage fields patch (#303)
* fix: manually running custom homepage fields patch

before this the patch was being skipped as while installing webshop the
"Website Items" table would be created and thus the install script
skipped all patches. but this patch is mandatory to be executed as
without this patch the homepage overrides made for webshop misbehave.

* fix: remove condition which checked for "Website Item" table

create_website_items handles the check

* fix: remove condition which checked for "Website Item" table

re: fix patch name